fix: array change dtype

+++ ivy/functional/frontends/jax/numpy/fft.py
@@ -31,5 +31,5 @@ def fft(a, n=None, axis=-1, norm=None):
 
 @to_ivy_arrays_and_back
 def ifftn(a, s=None, axes=None, norm=None):
-    a = ivy.asarray(a, dtype=ivy.complex128)
+    a = ivy.asarray(a, dtype=ivy.complex64)
     return ivy.ifftn(a, s=s, axes=axes, norm=norm)
